pyproject.toml
[project]
name = "pub-python-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adlfs>=2026.4.0",
    "alembic>=1.18.4",
    "fastapi>=0.135.2",
    "fsspec>=2026.3.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
